requirements.txt
cryptography==41.0.7
colorama==0.4.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
